requirements.txt
google-cloud-bigquery==1.16.0
